Rename parquet-go -> parquet
